Summary of additional changes to 1D transport
 Jan 18 & 19 2025

I decided to remove the Mass pulse from the 1D transport app

It presented some logistical issues for display given that the concentration starts as infinite so there is never any good way of establishing the graph axes.

In addition to that it does not lend itself to being presented to novices who want to "see" the mass spread over a volume

I am going to add a 3D spreading app soon - I will start on that as soon as I finish the 1D app. That will include a mass pulse

I moved the continuous source to be the first computational page because the volume source is an extension of the continuous source in that it is a superposition of two continuous sources

As I mentioned yesterday I adjusted parameters to represent a laboratory column that is 1 meter long and has the area and inflow rate necessary to establish a specific discharge of 0.0001 m/s then when the porosity varies from 0.01 to 0.6 the average travel time along the column ranges from 1.7 to 100 minutes

I allowed alpha to range 0.001 to 0.1 because 1 is the full length of the column and so is not reasonable for this problem

I removed the option to plot adv and ADE separately - so it always shows both - I think this is more useful from an educational perspective

I made similar changes for the finite pulse

Also for the finite pulse, I set the input C is to 1000g/m^3 and the Darcy discharge is 2.0268E-7 m^3/s , so each second introduces 2.0268E-4 g/s THUS given that the travel time along the length of the column can range from 100 to 6000 s and input of 0.020268 g would require 100 seconds at the shortest travel time while 1.2161 g would require 6000 seconds at the longest travel time so allow g to range from 0.001 to 0.1

The input was still in terms of Co so I changed that to mass and use the mass to calculate duration

I have not had time to check this yet but will do so when I get to work on it again which should be in a day or so