Eileen Poeter
I made a very minor edit. I looked at the docs that Thomas and Francesca sent. My conclusion is that we need to change the data sets that we are using for the examples. After looking at what we have and the information Francesca sent I have decided that I need you, Thomas, to select the data sets that will be used based on their known geology so that the use a confined leaky setting for Hantush and an unconfined setting for Neuman. Once you have selected suitable data sets, please also change the data in the py codes and the associate text naming the aquifer and describing the thicknesses and pumping rates and so on. I am not familiar with these data sets at all so I do not feel comfortable making these decisions and edits. After you have made these adjustments, please let me know and I will promptly proceed with the wording and format editing. I so appreciate the thought you are both putting into this,